repository cleanup and refactor: modified the the structure added a backend to logically separate code
